add missing staticrepo index file
gitignore was keeping it off the repository, and we hadn't noticed.

The file weighs 1.6MB, but it's a great way to test with a real index
with the right size and signature. Moreover, we likely won't be updating
this file often - perhaps once every couple of years, as the index
format progresses.

Most importantly, the file is not part of any importable Go package, so
it should never find its way to a vendor folder or a module archive.